QQ QQ QQ

+++ src/server/game/Spells/Spell.cpp
@@ -3924,7 +3924,7 @@ void Spell::SendSpellStart()
         data.append(m_caster->GetPackGUID());
 
     data.append(m_caster->GetPackGUID());
-    data << uint16(m_cast_count);                            // pending spell cast?
+    data << uint8(m_cast_count);                            // pending spell cast?
     data << uint32(m_spellInfo->Id);                        // spellId
     data << uint32(castFlags);                              // cast flags
     data << uint32(m_timer);                                // delay?
